More gui

+++ 3DModelViewer/imgui/imgui.h
@@ -1043,7 +1043,7 @@ struct ImDrawCmd
 
 // Vertex index (override with, e.g. '#define ImDrawIdx unsigned int' in ImConfig)
 #ifndef ImDrawIdx
-typedef unsigned short ImDrawIdx;
+typedef unsigned int ImDrawIdx;
 #endif
 
 // Vertex layout
